Rename files as ask instead of prompt
