+++ lib/middleware/template-ask-continue.js
